Move Join fragments to fragements.join
