+++ babble/src/main/java/io/mosaicnetworks/babble/fragments/join/WebRTCJoinGroupFragment.java
@@ -22,7 +22,7 @@
  * SOFTWARE.
  */
 
-package io.mosaicnetworks.babble.configure.webrtc;
+package io.mosaicnetworks.babble.fragments.join;
 
 import android.app.ProgressDialog;
 import android.content.Context;
